pyproject.toml
[tool.poetry]
name = "dailynator"
version = "1.0.0"
description = "Slack bot for handling with daily meetings in ease"
authors = ["Drugsosos <44712637+Drugsosos@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
slack-bolt = "^1.14.3"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.1"
unsync = "^1.4.0"
APScheduler = "^3.9.1"
websockets = "^10.3"
redis = "^4.2.0rc1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
